Correção dos tipos de id's
Mudança de int para long

+++ app/src/main/java/com/adht/android/medicontrol/alarme/negocio/AlarmeServices.java
@@ -34,7 +34,7 @@ public class AlarmeServices {
         dao.deletar(idAlarme);
     }
 
-    public Alarme getAlarme(int idAlarme){
+    public Alarme getAlarme(long idAlarme){
         Alarme alarme = new Alarme();
         AlarmeDAOSQLite dao = new AlarmeDAOSQLite();
         try {
